Update Graph Continuously (not Asynchronous) and Enemy Controller Patches
AstarPath cannot run on a thread that isn't main, so I set it to run in main on set time intervals (as to limit how much the function boggs down the game). As a result, it's updates are slightly delayed, but it doesn't halt progress as frequently. I found out the Enemy Controller was producing errors and added some if statements to avoid them.